+++ flask_restly/decorator/delete.py
@@ -2,7 +2,7 @@ from flask_restly._storage import append_mapping
 from functools import wraps
 
 
-def delete(path, queued=False, serializer=None):
+def delete(path, queued=False, serialize=None):
     def decorator(func):
         @wraps(func)
         def wrapper(*args, **kwargs):
@@ -10,7 +10,7 @@ def delete(path, queued=False, serializer=None):
 
             return ('', 204) if result is None else (result, 200 if not queued else 202)
 
-        append_mapping(wrapper, path, serializer, 'DELETE')
+        append_mapping(wrapper, path, serialize, 'DELETE')
 
         return wrapper