+++ flask_restly/decorator/get.py
@@ -2,7 +2,7 @@ from flask_restly._storage import append_mapping
 from functools import wraps
 
 
-def get(path, serializer=None):
+def get(path, serialize=None):
     def decorator(func):
         @wraps(func)
         def wrapper(*args, **kwargs):
@@ -10,7 +10,7 @@ def get(path, serializer=None):
 
             return response, 200
 
-        append_mapping(wrapper, path, serializer, 'GET')
+        append_mapping(wrapper, path, serialize, 'GET')
 
         return wrapper